+++ src/components/Card/index.js
@@ -1,9 +1,22 @@
 import Card from './Card';
+import CardBody from './CardBody';
+import CardFooter from './CardFooter';
+import CardHeader from './CardHeader';
+import CardImage from './CardImage';
+import CardSubtitle from './CardSubtitle';
+import CardTitle from './CardTitle';
+
 import { usePlugin, installPlugin } from '@/utils';
 
 const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Card);
+    installPlugin(Vue, CardBody);
+    installPlugin(Vue, CardFooter);
+    installPlugin(Vue, CardHeader);
+    installPlugin(Vue, CardImage);
+    installPlugin(Vue, CardSubtitle);
+    installPlugin(Vue, CardTitle);
   },
 };
 
@@ -13,4 +26,10 @@ export default VuePlugin;
 
 export {
     Card,
+    CardBody,
+    CardFooter,
+    CardHeader,
+    CardImage,
+    CardSubtitle,
+    CardTitle,
 };